fix(frontend): Ensure the Forecase image overlays on map correctly
The Image overlay was not displaying correctly on the map, it should
have covered the whole of Europe, but only covered the AOI set for
the individual user. This isn't the best of solutions as if there
were a forecast for the US say, then this wouldn't work. I'm not
sure how we could make this work with the current info available.

The actual fix is to define the bbox for Europe and since the data
is in WGS84 and not WGS83 Mercator, which mapbox maps use, we also
need to add the COORDINATE SYSTEM, the raster image comes in, namely
LNG/LAT.