fix(frontend): Remove unused code
Since we are defining the BBOX to be for the whole of Europe, the
code to calculate the bbox from the user's selected AOI is no longer
relevant. I'm still not happy with this as the BBOX is hard-coded
for Europe, if we say, wanted it to work for the US, then this code
would not work. However, I would say that the data should be in
WGS84 Web Mercator (EPSG:3587) projection, not in WGS84 (EPSG:4326)
as it currently is as mapbox works with Web Mercator, so changing
it in code is not great. Also, the true fix would be to provide a
TMS that the mapbox code would work with, rather than shoe-horn in
WMS requests.