fix(frontend): Prevent popup crashing, some light styling
For some reason, `object.properties` passed to the popup is now `object.properties.properties`,
no idea why. This fixes (and addes some elvis operators to prevent crashing in future), but is
indicative of a larger issue (how did that happen).

Also added in a break tag to format the lat/lon values more nicely.